+++ CHANGELOG.md
@@ -29,3 +29,8 @@ All notable changes to the "html2ft" extension will be documented in this file.
 ## 0.7.0
 
 - Improve code formatting
+
+## 0.8.0
+
+- Correctly parse alpinejs attrs like `x-transition.opacity.duration.600ms`
+- Group kw args into 1 dictionary `Div(**{'x-transition.opacity.duration.600ms': True},  **{'x-transition.scale.origin.top': True})` becomes `Div(**{'x-transition.opacity.duration.600ms': True, 'x-transition.scale.origin.top': True})`